refactor(favorite): move add favorite logic from UserController to FavoriteController
